team10/backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
click==8.2.1
fastapi==0.116.1
h11==0.16.0
httptools==0.6.4
idna==3.10
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-core
python-dotenv==1.1.1
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
pandas
geopy==2.4.1
pycountry
rapidfuzz
pydantic[email]
python-dotenv==1.1.1
python-dateutil